python_server/src/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
grpcio==1.37.1
grpcio-tools==1.37.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mmh3==3.0.0
numpy==1.22.3
opencv-python==4.5.5.64
pandas==1.4.2
pdqhash==0.2.2
Pillow==9.1.0
protobuf==3.20.1
pymilvus==2.0.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.16.6
six==1.16.0
tabulate==0.8.9
timm==0.5.4
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
towhee==0.6.1
tqdm==4.64.0
typing-extensions==4.2.0
ujson==5.1.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
